+++ src/guards/index.ts
@@ -1,0 +1,1 @@
+export { default as isTypedArray } from './isTypedArray'