+++ CHANGELOG.md
@@ -1,1 +1,6 @@
 # Changelog
+
+## [1.1.0] 03-16-2025
+
+- Add `uniqueTA` to optimally filter duplicates.
+- Add `isTypedArray` as quick and small guard function.